check if self delete

+++ internal/handlers/users.go
@@ -2,6 +2,7 @@ package handlers
 
 import (
 	"encoding/json"
+	"errors"
 	"net/http"
 	db "receipt-wrangler/api/internal/database"
 	"receipt-wrangler/api/internal/models"
@@ -217,6 +218,11 @@ func DeleteUser(w http.ResponseWriter, r *http.Request) {
 		ResponseType: "",
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
 			id := chi.URLParam(r, "id")
+			token := utils.GetJWT(r)
+			if utils.UintToString(token.UserId) == id {
+				return 500, errors.New("user cannot delete itself")
+			}
+
 			err := services.DeleteUser(id)
 			if err != nil {
 				return 500, err
